refactored RenderDates to include the filter box
