--- frontend/components/date_type_filter/filter/filter.jsx
@@ -1,34 +1,0 @@
-import React, { useEffect, useState } from 'react'
-
-
-export const DateFilter = ({dates}) => {
-    const [checkedBox, setCheckedBox] = useState({
-        costFilter: false, 
-        dateNumberFilter: false,
-        locationFilter: false,
-        dateTypeFilter: false
-    })
-    console.log(dates)
-    return (
-        <>
-            <h1>HASDFASDF</h1>
-        </>
-    )
-}
-
-
-const costFilter = (filteredDates) => {
-    
-}
-
-const dateNumberFilter = (filteredDates) => {
-
-}
-
-const locationFilter = (filteredDates) => {
-
-}
-
-const dateTypeFilter = (filteredDates) => {
-
-}
